requirements.txt
SpeechRecognition
pipwin
pyaudio
gTTS
google-generativeai
python-dotenv
streamlit
-e .



# Its solved you need Python 3.10 to work close this issue now.